Added ProductHistory filter; More Order filters; Plus tests;

+++ Application/Api/Products/Delete/Handlers/DeleteProductHandler.cs
@@ -9,10 +9,12 @@ namespace Application.Api.Products.Delete.Handlers;
 public class DeleteProductHandler : IRequestHandler<DeleteProductCommand, OneOf<DeleteProductResult, List<PlainApplicationError>>>
 {
     private readonly IProductRepository _productRepository;
+    private readonly IProductHistoryRepository _productHistoryRepository;
 
-    public DeleteProductHandler(IProductRepository productRepository)
+    public DeleteProductHandler(IProductRepository productRepository, IProductHistoryRepository productHistoryRepository)
     {
         _productRepository = productRepository;
+        _productHistoryRepository = productHistoryRepository;
     }
 
     public async Task<OneOf<DeleteProductResult, List<PlainApplicationError>>> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
@@ -29,6 +31,23 @@ public class DeleteProductHandler : IRequestHandler<DeleteProductCommand, OneOf<
             };
         }
 
+        var latestProductHistory = await _productHistoryRepository.GetLatestByProductIdAsync(product.Id);
+        if (latestProductHistory is null)
+        {
+            return new List<PlainApplicationError>() {
+                new PlainApplicationError(
+                    message: $"Product of Id \"{request.Id}\" lacks valid ProductHistory.",
+                    path: ["_"],
+                    code: ValidationErrorCodes.IntegrityError
+                )
+            };
+        }
+
+        // Invalidate old history
+        latestProductHistory.Invalidate();
+        await _productHistoryRepository.UpdateAsync(latestProductHistory);
+
+        // Delete product
         await _productRepository.DeleteByIdAsync(product.Id);
 
         var result = new DeleteProductResult();
